A new font is added
