+++ src/components/PhotoSlider.jsx
@@ -1,0 +1,27 @@
+import Slider from "react-slick";
+
+function PhotoSlider() {
+  const settings = {
+    dots: true,
+    infinite: true,
+    speed: 500,
+    slidesToShow: 1,
+    slidesToScroll: 1,
+  };
+
+  return (
+    <div>
+      <h2> Single Item</h2>
+      <Slider {...settings}>
+        <div>
+          <h3>1</h3>
+        </div>
+        <div>
+          <h3>2</h3>
+        </div>
+      </Slider>
+    </div>
+  );
+}
+
+export default PhotoSlider;
