+++ src/css-blocks/header/__name/header__name.css
@@ -1,0 +1,7 @@
+.header__name {
+  color: white;
+  margin: 16px auto auto auto;
+  font-family: Mullish, arial, sans-serif;
+  font-weight: 500;
+  margin-bottom: 10px;
+}
