+++ src/css-blocks/header/__photo/header__photo.css
@@ -2,5 +2,6 @@
   max-width: 200px;
   max-height: 300px;
   border-radius: 1rem;
-  margin-left: 3rem;
+  margin: auto;
+  box-shadow: rgb(34, 34, 34) 2px 2px 2px 2px;
 }
