+++ src/css-blocks/header/__subname/header__subname.css
@@ -1,0 +1,7 @@
+.header__subname {
+  color: white;
+  margin: 0 auto auto auto;
+  font-family: Mullish, arial, sans-serif;
+  font-weight: 500;
+  margin-bottom: 10px;
+}
