+++ src/css-blocks/header/__top/header__top.css
@@ -1,0 +1,24 @@
+.header__top {
+  font-family: Mullish, arial, sans-serif;
+  font-weight: 500;
+  max-width: min-content;
+  color: #fff;
+  overflow: hidden; /* Ensures the content is not revealed until the animation */
+  white-space: nowrap; /* Keeps the content on a single line */
+  border-right: .15em solid rgba(10, 10, 10, 0); /* The typwriter cursor */
+  margin: 6px auto 10px auto; /* Gives that scrolling effect as the typing happens */
+  animation: 
+    typing 3.5s steps(30, end),
+    blink-caret .5s step-end 3;
+}
+
+/* The typing effect */
+@keyframes typing {
+  from { width: 0 }
+  to { width: 100% }
+}
+
+@keyframes blink-caret {
+  from, to { border-color: transparent }
+  50% { border-color: rgb(255, 255, 255) }
+}
