--- src/css-blocks/header/name/header__name.css
@@ -1,4 +1,0 @@
-.header__name {
-  color: white;
-  margin: auto;
-}
