--- src/css-blocks/header/top/header__top.css
@@ -1,4 +1,0 @@
-.header__top {
-  color: white;
-  margin: auto auto 40px auto ;
-}
