+++ src/index.css
@@ -1,15 +1,18 @@
-@import url('./vendor/normalize.css');
-@import url('./css-blocks/app/app.css');
-@import url('./css-blocks/header/header.css');
-@import url('./css-blocks/header/top/header__top.css');
-@import url('./css-blocks/header/name/header__name.css');
-@import url('./css-blocks/header/photo/header__photo.css');
-@import url('./css-blocks//header/container/header__container.css');
-@import url('./css-blocks/app/__color/_type/app__color_type_cyan-pink.css');
-@import url('./css-blocks/app/__color/_type/app__color_type_cyan-purple.css');
-@import url('./css-blocks/app/__color/_type/app__color_type_cyan-yellow.css');
-@import url('./css-blocks/app/__color/_type/app__color_type_blue-red.css');
-@import url('./css-blocks/app/__color/_type/app__color_type_green-blue.css');
-@import url('./css-blocks/app/__color/_type/app__color_type_pink-yellow.css');
-@import url('./css-blocks/app/__color/_type/app__color_type_red-green.css');
-@import url('./css-blocks/app/__color/_type/app__color_type_active.css');
+@import url("./vendor/normalize.css");
+@import url("./vendor/fonts/fonts.css");
+@import url("./css-blocks/app/app.css");
+@import url("./css-blocks/header/header.css");
+@import url("./css-blocks/header/__top/header__top.css");
+@import url("./css-blocks/header/__name/header__name.css");
+@import url("./css-blocks/header/__photo/header__photo.css");
+@import url("./css-blocks//header/__container/header__container.css");
+@import url("./css-blocks/app/__color/_type/app__color_type_cyan-pink.css");
+@import url("./css-blocks/app/__color/_type/app__color_type_cyan-purple.css");
+@import url("./css-blocks/app/__color/_type/app__color_type_cyan-yellow.css");
+@import url("./css-blocks/app/__color/_type/app__color_type_blue-red.css");
+@import url("./css-blocks/app/__color/_type/app__color_type_green-blue.css");
+@import url("./css-blocks/app/__color/_type/app__color_type_pink-yellow.css");
+@import url("./css-blocks/app/__color/_type/app__color_type_red-green.css");
+@import url("./css-blocks/app/__color/_type/app__color_type_active.css");
+@import url("./css-blocks/header/__subname/header__subname.css");
+
